requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
requests
pytest
pytest-django
drf-yasg